vendor: move github.com/huin/goupnp out of vendor
Because github.com/huin/goupnp was modified for specific use by this
repo, it became difficult to correctly vendor this using a tool. To
avoid this issue, github.com/huin/goupnp was taken out of the vendor dir
and included as a packge in the repo.
