+++ goupnp/ssdp/registry.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/huin/goupnp/httpu"
+	"github.com/NebulousLabs/go-upnp/goupnp/httpu"
 )
 
 const (
